Add lazy session support when using the :expires_at option
In Rails, the ActionController::Session::AbstractStore does not
allow for lazy session loading. It will always create a new
session. This branch overrides the #call method to not take
:expires_at options into account when deciding whether to start
a session.

The difference is that the expires time of a cookie will only
get updated when the cookie is accessed in Rails.